Remove the faking of the TTY in worker processes
