+++ lib/fork.js
@@ -12,13 +12,6 @@ if (fs.realpathSync(__filename) !== __filename) {
 // the presence of this variable allows it to require this one instead
 const AVA_PATH = path.resolve(__dirname, '..');
 
-const describeTTY = tty => ({
-	colorDepth: tty.getColorDepth ? tty.getColorDepth() : undefined,
-	hasColors: typeof tty.hasColors === 'function',
-	columns: tty.columns || 80,
-	rows: tty.rows
-});
-
 const workerPath = require.resolve('./worker/subprocess');
 
 module.exports = (file, opts, execArgv) => {
@@ -34,10 +27,6 @@ module.exports = (file, opts, execArgv) => {
 	opts = {
 		file,
 		baseDir: process.cwd(),
-		tty: {
-			stderr: process.stderr.isTTY ? describeTTY(process.stderr) : false,
-			stdout: process.stdout.isTTY ? describeTTY(process.stdout) : false
-		},
 		...opts
 	};
