+++ lib/worker/subprocess.js
@@ -9,7 +9,6 @@ ipc.send({type: 'ready-for-options'});
 ipc.options.then(options => {
 	require('./options').set(options);
 	require('../chalk').set(options.chalkOptions);
-	require('./fake-tty'); // eslint-disable-line import/no-unassigned-import
 
 	const babelManager = require('../babel-manager');
 	const nowAndTimers = require('../now-and-timers');
