--- test/fixture/tty/color-disabled.js
@@ -1,9 +1,0 @@
-const test = require('../../..');
-
-const assertNoColor = (t, stream) => {
-	t.true(stream.isTTY);
-	t.is(stream.getColorDepth(), 1);
-};
-
-test('stderr does not support color', assertNoColor, process.stderr);
-test('stdout does not support color', assertNoColor, process.stdout);
