--- test/fixture/tty/color-enabled.js
@@ -1,11 +1,0 @@
-const test = require('../../..');
-
-const assertColor = (t, stream) => {
-	t.true(stream.isTTY);
-	const colorDepth = stream.getColorDepth();
-	t.is(typeof colorDepth, 'number');
-	t.true(colorDepth > 1);
-};
-
-test('stderr supports color', assertColor, process.stderr);
-test('stdout supports color', assertColor, process.stdout);
